Split utils
