+++ station/tests/test_orbitdb.py
@@ -7,7 +7,7 @@ from shutil import copy, rmtree
 # It must be set before import "utils"
 environ["AQUARIUS_CONFIG_DIR"] = "tests/config"
 
-from utils import CONFIG_DIRECTORY
+from utils.globalvars import CONFIG_DIRECTORY
 
 from orbitdb import OrbitDatabase
