+++ station/utils/editwatcher.py
@@ -1,0 +1,24 @@
+class EditWatcher:
+    def __init__(self, obj):
+        object.__setattr__(self, "_obj", obj)
+        object.__setattr__(self, "_edited", False)
+
+    def __getattr__(self, key):
+        return getattr(self._obj, key)
+
+    def __setattr__(self, name, value):
+        object.__setattr__(self, "_edited", True)
+        return setattr(self._obj, name, value)
+
+    def __getitem__(self, key):
+        return self._obj[key]
+
+    def __setitem__(self, key, value):
+        object.__setattr__(self, "_edited", True)
+        self._obj[key] = value
+
+    def is_edited(self):
+        return self._edited
+
+    def get_wrapped(self):
+        return self._obj
